requirements.txt
google-adk==1.27.4
google-cloud-firestore==2.19.0
google-cloud-aiplatform>=1.38.0
python-dotenv>=1.1.0
fastmcp==3.1.1
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0
